BOSS: merge ambalink fixed from kernel 3.8.
Including,
1. Inserted AMBAFS_DMSG for debug mode.
2. Fixed sqlite test.
3. Fixed to support append mode.
4. Fixed sqlite_load_test.
5. Fixed virt-to-phys address translation for ARM11 Linux.
6. Refined. Use ambalink_lkm_virt_to_phys() macro for address
   translation.
7. Refined: removed unused defined.
8. Fixed compiled warnings.

+++ drivers/rpmsg/virtio_rpmsg_bus.c
@@ -592,7 +592,7 @@ static void *get_a_tx_buf(struct virtproc_info *vrp, int *idx)
 {
 	unsigned int len;
 	void *ret;
-	unsigned int flags;
+	unsigned long flags;
 
 	/* support multiple concurrent senders */
 	rpmsg_tx_lock(&vrp->tx_lock, flags);
@@ -633,7 +633,7 @@ static void *get_a_tx_buf(struct virtproc_info *vrp, int *idx)
  */
 static void rpmsg_upref_sleepers(struct virtproc_info *vrp)
 {
-	unsigned int flags;
+	unsigned long flags;
 
 	/* support multiple concurrent senders */
 	rpmsg_tx_lock(&vrp->tx_lock, flags);
@@ -662,7 +662,7 @@ static void rpmsg_upref_sleepers(struct virtproc_info *vrp)
  */
 static void rpmsg_downref_sleepers(struct virtproc_info *vrp)
 {
-	unsigned int flags;
+	unsigned long flags;
 
 	/* support multiple concurrent senders */
 	rpmsg_tx_lock(&vrp->tx_lock, flags);
@@ -717,7 +717,7 @@ int rpmsg_send_offchannel_raw(struct rpmsg_channel *rpdev, u32 src, u32 dst,
 	struct scatterlist sg;
 	struct rpmsg_hdr *msg;
 	int err, idx;
-	unsigned int flags;
+	unsigned long flags;
 
 #if RPMSG_DEBUG
 	unsigned int prev_time, current_time, diff;
